Move buildItemTableDisplayingString to itemUtils
