+++ src/utils/itemUtils.ts
@@ -1,4 +1,6 @@
 import type { ItemLocationKey, ItemLocation } from "../types/itemTypes"
+import { getPaginationOffsetStrings } from "./appUtils"
+import { ITEM_PAGINATION_BATCH_SIZE } from "../config/constants"
 
 export const itemAvailableIds = ["status:a", "status:o"]
 
@@ -25,6 +27,33 @@ export const locationEndpointsMap: Record<ItemLocationKey, string> = {
 // Extract location key from the location label in the API response
 export function locationLabelToKey(label: string): ItemLocationKey {
   return label.replace(/SASB/, "Schwarzman").split(" ")[0] as ItemLocationKey
+}
+
+// Build the heading above the Item Table in the Bib Page
+// based on pagination values
+export const buildItemTableDisplayingString = (
+  page: number,
+  totalResults: number,
+  viewAllItems = false
+) => {
+  const isPlural = totalResults > 1
+  const totalString = totalResults.toLocaleString()
+
+  if (viewAllItems || totalResults <= ITEM_PAGINATION_BATCH_SIZE) {
+    return isPlural
+      ? `Displaying all ${totalString} items`
+      : "Displaying 1 item"
+  }
+
+  const [resultsStart, resultsEnd] = getPaginationOffsetStrings(
+    page,
+    totalResults,
+    ITEM_PAGINATION_BATCH_SIZE
+  )
+
+  return `Displaying ${resultsStart}-${resultsEnd} of ${totalResults.toLocaleString()} item${
+    isPlural ? "s" : ""
+  }`
 }
 
 /**
