+++ src/utils/utilsTests/bibUtils.test.ts
@@ -1,9 +1,4 @@
-import {
-  standardizeBibId,
-  isNyplBibID,
-  getBibQueryString,
-  buildItemTableDisplayingString,
-} from "../bibUtils"
+import { standardizeBibId, isNyplBibID, getBibQueryString } from "../bibUtils"
 
 describe("bibUtils", () => {
   describe("standardizeBibId", () => {
@@ -50,31 +45,6 @@ describe("bibUtils", () => {
     it("returns the correct query string with a bib ID and query params", () => {
       expect(getBibQueryString({ id: "b12082323", item_page: 5 })).toBe(
         "?items_size=20&items_from=80&item_page=5&id=b12082323&merge_checkin_card_items=true"
-      )
-    })
-  })
-  describe("buildItemTableDisplayingString", () => {
-    it("returns the correct item table heading when there is one item", () => {
-      expect(buildItemTableDisplayingString(1, 1)).toBe("Displaying 1 item")
-    })
-    it("returns the correct item table heading when the total number is greater than 1 but less than the pagination limit", () => {
-      expect(buildItemTableDisplayingString(1, 2)).toBe(
-        "Displaying all 2 items"
-      )
-    })
-    it("returns the correct item table heading for first page when there are many items", () => {
-      expect(buildItemTableDisplayingString(1, 300)).toBe(
-        "Displaying 1-20 of 300 items"
-      )
-    })
-    it("returns the correct item table heading for pages greater than 1 when there are many items", () => {
-      expect(buildItemTableDisplayingString(5, 300)).toBe(
-        "Displaying 81-100 of 300 items"
-      )
-    })
-    it("returns the correct item table heading for when view all items is enabled", () => {
-      expect(buildItemTableDisplayingString(1, 300, true)).toBe(
-        "Displaying all 300 items"
       )
     })
   })
